AddNotification AddViewFlag Change lng and lat to double

+++ TrackerApi/Models/Location.cs
@@ -14,9 +14,9 @@ namespace TrackerApi.Models
         [DataType(DataType.Text, ErrorMessage = "Please enter valid Name")]
         public string Name { get; set; }
         [Required]
-        public int Log { get; set; }
+        public double Log { get; set; }
         [Required]
-        public int Lat { get; set; }
+        public double Lat { get; set; }
 
         public bool viewFlag { get; set; }
